testing with directories, add package where libraries are tested
